[state] Replace std mutex with parking_lot and fix unwrap
