+++ state/state-tree/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2018"
 
 [dependencies]
 anyhow = "1.0.38"
+parking_lot = "0.11.1"
 starcoin-types = {path = "../../types"}
 starcoin-crypto = { path = "../../commons/crypto"}
 serde = { version = "1.0.123" }
